[TST] Add test for peaks2maps (#328)
* Add test for peaks2maps.

Also, remove peaks2maps submodule, moving function into meta utils submodule.

Clean up data dir handling as well.

* Install tensorflow...

* Trigger rebuild.

* Update black and re-run.

* Update test_meta_mkda.py

* Drop negcore test.

* Split KDA test.
